Parse Fix
This is causing problems when trying to bring this device up as a template in the IDE as well as when using addChildDevice("smartthings/testing", "Simulated Switch", deviceId, null, [label: switchLabel]) to create a virtual device from a SmartApp. It's a virtual device so it shouldn't be getting anything to parse anyway.